+++ android-appcoins-billing/src/debug/AndroidManifest.xml
@@ -1,11 +1,8 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
     package="com.appcoins.billing.sdk">
 
   <uses-permission android:name="android.permission.WAKE_LOCK"/>
   <uses-permission android:name="com.appcoins.BILLING"/>
-  <uses-permission android:name="android.permission.QUERY_ALL_PACKAGES"
-      tools:ignore="QueryAllPackagesPermission" />
 
   <queries>
     <intent>
@@ -50,6 +47,5 @@
           </intent-filter>
         </activity>
     <service android:name="com.appcoins.sdk.billing.service.WebPaymentWebSocketService"/>
-    <service android:name="com.appcoins.sdk.billing.service.BillingLifecycleService" />
   </application>
 </manifest>
