+++ android-appcoins-billing/src/main/java/com/appcoins/sdk/billing/managers/BillingLifecycleManager.kt
@@ -2,20 +2,38 @@ package com.appcoins.sdk.billing.managers
 
 import android.content.Context
 import android.content.Intent
-import com.appcoins.sdk.billing.service.BillingLifecycleService
+import android.content.IntentFilter
+import android.util.Log
+import com.appcoins.sdk.billing.payflow.PayflowManager
+import com.appcoins.sdk.billing.receivers.AppInstallationReceiver
 
 object BillingLifecycleManager {
 
+    private const val PACKAGE_SCHEME = "package"
+    private val appInstallationReceiver by lazy { AppInstallationReceiver() }
+
     @JvmStatic
-    fun initializeBillingLifecycleService(context: Context) {
-        val intent = Intent(context, BillingLifecycleService::class.java)
-        context.applicationContext.startService(intent)
+    fun setupBillingService(context: Context) {
+        Thread {
+            AttributionManager.getAttributionForUser()
+            PayflowManager.getPayflowPriorityAsync()
+            val receiverIntentFilter = IntentFilter()
+            receiverIntentFilter.addAction(Intent.ACTION_PACKAGE_ADDED)
+            receiverIntentFilter.addAction(Intent.ACTION_PACKAGE_REMOVED)
+            receiverIntentFilter.addDataScheme(PACKAGE_SCHEME)
+            context.applicationContext.registerReceiver(
+                appInstallationReceiver,
+                receiverIntentFilter
+            )
+        }.start()
     }
 
     @JvmStatic
-    fun stopBillingLifecycleService(context: Context) {
-        context.applicationContext.stopService(
-            Intent(context, BillingLifecycleService::class.java)
-        )
+    fun finishBillingService(context: Context) {
+        try {
+            context.unregisterReceiver(appInstallationReceiver)
+        } catch (e: Exception) {
+            Log.e("BillingLifecycleManager", "Failed to unregister AppInstallationReceiver: $e")
+        }
     }
 }
